Boot animation script
Inspired by Omni's and Lineage's boot animation script, I wanted a way for
not only new maintainers to get a boot animation that completely fits their
device but also hook up the unofficial folks. Unofficial maintainers are often
left out in the dark with boot animations that just don't fit their devices.

A lot of times devs assume that one size fits all but that's not usually the
case unless your boot animation is very small and centered. If you have a
large boot animation that reaches the edges you start to noice that one size
does not fit all. This looks janky.

This script takes care of that by allowing us to quickly resize our current
boot animation to EXACTLY the size we need it to be.

Also unlike Omni's or Lineage's script, this one can be launched just like we
do repopick or the roomservice so is buildbox friendly :p

For unofficial folks or for new maintainers, the ideal use case is to launch
the script in the same way you do repopick

- . build/envsetup.sh
- bootanimation width heigh (For example, bootanimation 1920 1200)
- The script runs and replaces the bootanimation in vendor/du for their use
- Is that simple!

Once is there they can compile, use it or grab and save it for future use

Change-Id: Idd32058ba4f4e2958fcf7a3e6f6ed44e856388d1